perf: vector columns use storage plain
Based on prior benchmarking, using storage plain (rather than
extended/external) for vector columns performs much better. We now set the
storage on the column to plain if the dimensions are <= 2000 for both existing
and future vectorizers.

NOTE: If you have existing vectorizers where the dimensions are less than 2001,
we will set the storage to plain, but this will not take effect until a table
rewrite happens. This can be done with `CLUSTER` or `VACUUM FULL` and may take
a while depending on the size of the table.